requirements.txt
PyMuPDF
opencv-python
numpy
pandas
python-dotenv
Pillow
requests